requirements.txt
requests==2.*
yt-dlp==2025.*
python-dotenv==1.*
azure-storage-blob==12.*